Create build.c and do an entire 1 file upload
"How did you ever do it? It seems impossible!" "Well, I didn't give up. They said I couldn't, but I defied the odds and-" "Dude, it's one file. Not so epic." "Still, it's something." "No, it isn't." "Yes it is, it's something. It's not a lot, but it's definitely something." "Oh, screw you."